src/requirements.txt
fastapi==0.121.3
faiss-cpu==1.13.0
requests==2.32.5
dotenv==0.9.9
pypdf==6.3.0
pydantic==2.12.4
uvicorn==0.38.0
numpy==2.3.5